✨ feat: Improve SimpleRotationAnimation by adding a speedUp method
This commit adds a new method `speedUp` to the `SimpleRotationAnimationState` class, which allows for changing the duration of the animation and repeating it. The duration is now set to 1 second instead of 2 seconds. Additionally, a `GlobalKey` is added to the `_MyHomePageState` class, which is used to access the `SimpleRotationAnimationState` instance and call `speedUp` with an updated duration value each time the button is pressed. The text displaying the number of button presses is changed to display the duration value instead of the button press count.