✨ feat: Add ability to dynamically scale SimpleRotationAnimation
This commit adds the ability to dynamically scale the SimpleRotationAnimation widget by introducing a `scaleUp` method in the `SimpleRotationAnimationState` class. The method takes a new size as an argument and updates the `_size` variable accordingly using `setState`. The size is then used to update the width and height of the Container within the widget.

In addition, the commit also modifies the `_MyHomePageState` class to call the new `scaleUp` method with a calculated size based on the current duration of the animation.

This commit improves the functionality of the SimpleRotationAnimation widget by allowing it to be dynamically scaled at runtime, which can enhance its visual appeal and provide greater flexibility for UI designers and developers.