✨ feat: Add ability to change scale in SimpleRotationAnimation
This commit adds the ability to change the scale of the rotating container in SimpleRotationAnimation. The `changeScale()` method is introduced, which toggles the scale between 1.0 and 3.0 using `setState()`. An `AnimatedScale` widget is used to animate the scaling transition with a duration of 2 seconds.

Additionally, the `scaleUp()` method previously used to increase the size of the container has been modified to take only one argument instead of two. This argument now represents the percentage by which the size will be scaled up.

Finally, in main.dart, the call to `scaleUp()` has been replaced with the call to `changeScale()`.